pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "occp"
version = "0.9.0"
description = "OpenCloud Control Plane – Agent Control Plane with Verified Autonomy Pipeline"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12,<3.14"  # 3.14 has intermittent SQLAlchemy C extension crashes
authors = [
    {name = "Azar Management Consulting", email = "azaroccp@gmail.com"},
]
keywords = ["agent", "control-plane", "orchestrator", "policy", "ai", "mcp", "governance"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "click>=8.1",
    "websockets>=14.0",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "aiosqlite>=0.20",
    "PyJWT>=2.10",
    "sqlalchemy[asyncio]>=2.0.36",
    "alembic>=1.14",
    "argon2-cffi>=23.1",
    "casbin>=1.40",
    "structlog>=24.4",
    "cryptography>=44.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
llm = ["anthropic>=0.40.0"]
openai = ["openai>=1.60"]
llm-all = ["occp[llm,openai]"]
postgres = ["asyncpg>=0.30"]
mcp = ["mcp>=1.6"]
observability = [
    "opentelemetry-api>=1.30",
    "opentelemetry-sdk>=1.30",
    "opentelemetry-exporter-otlp>=1.30",
]
presidio = [
    "presidio-analyzer>=2.2",
    "presidio-anonymizer>=2.2",
]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "httpx>=0.28",
    "ruff>=0.9",
    "mypy>=1.14",
]
all = ["occp[llm-all,postgres,mcp,observability,dev]"]

[project.scripts]
occp = "cli.main:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
include = [
    "orchestrator*", "policy_engine*", "cli*", "sdk*",
    "adapters*", "api*", "config*", "store*", "security*",
]

# ── Ruff (replaces black + isort + flake8 + bandit) ───────────────
[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E", "F", "W",     # pyflakes + pycodestyle
    "I",                # isort
    "N",                # pep8-naming
    "UP",               # pyupgrade (modern syntax)
    "S",                # bandit (security)
    "B",                # bugbear
    "SIM",              # simplify
    "TCH",              # type-checking imports
    "RUF",              # ruff-specific
    "ASYNC",            # async best practices
    "PTH",              # pathlib
    "T20",              # print statements
]
ignore = [
    "S101",             # assert in tests is fine
    "S603", "S607",     # subprocess — needed for sandbox executor
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S106", "T20"]
"scripts/**" = ["T20"]

[tool.ruff.format]
quote-style = "double"

# ── mypy ──────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Incremental — enable strict later
check_untyped_defs = true
no_implicit_optional = true

[[tool.mypy.overrides]]
module = ["casbin.*", "yaml.*"]
ignore_missing_imports = true
